Move the state machines to their own package to make it easier to get a grip on.
