+++ src/relstorage/storage/__init__.py
@@ -32,31 +32,39 @@ from persistent.timestamp import TimeStamp
 from ZODB import ConflictResolution
 from ZODB.blob import is_blob_record
 from ZODB.mvccadapter import HistoricalStorageAdapter
-from ZODB.POSException import ConflictError
+
 from ZODB.POSException import POSKeyError
 from ZODB.POSException import ReadConflictError
 from ZODB.POSException import ReadOnlyError
 from ZODB.POSException import StorageTransactionError
 from ZODB.POSException import Unsupported
+
 from ZODB.UndoLogCompatible import UndoLogCompatible
 from ZODB.utils import p64 as int64_to_8bytes
 from ZODB.utils import u64 as bytes8_to_int64
 from zope import interface
 from zope.interface import implementer
 
-from relstorage._compat import base64_decodebytes
+
 from relstorage._compat import base64_encodebytes
-from relstorage._compat import dumps
 from relstorage._compat import loads
-from relstorage._compat import OID_TID_MAP_TYPE
 from relstorage._compat import OID_SET_TYPE
-from relstorage._transaction_iterator import TransactionIterator
+
+
 from relstorage.blobhelper import BlobHelper
 from relstorage.cache import StorageCache
 from relstorage.cache.interfaces import CacheConsistencyError
 from relstorage.options import Options
 
-# pylint:disable=too-many-lines
+from .transaction_iterator import TransactionIterator
+from .tpc import NotInTransaction
+from .tpc.begin import HistoryFree
+from .tpc.begin import HistoryPreserving
+from .tpc.restore import Restore
+
+__all__ = [
+    'RelStorage',
+]
 
 log = logging.getLogger("relstorage")
 
@@ -69,12 +77,7 @@ abort_early = os.environ.get('RELSTORAGE_ABORT_EARLY')
 
 z64 = b'\0' * 8
 
-def _to_utf8(data):
-    if data is None:
-        return data
-    if isinstance(data, bytes):
-        return data
-    return data.encode("utf-8")
+# pylint:disable=too-many-lines
 
 class _DummyLock(object):
     # Enough like a lock for our purposes so we can switch
@@ -238,15 +241,12 @@ class RelStorage(UndoLogCompatible,
         elif options.blob_dir:
             self.blobhelper = BlobHelper(options=options, adapter=adapter)
 
-        if self._options.keep_history:
-            self._tpc_begin_factory = _HistoryPreservingBeginTPCState
-        else:
-            self._tpc_begin_factory = _HistoryFreeBeginTPCState
+        self._tpc_begin_factory = HistoryPreserving if self._options.keep_history else HistoryFree
 
         if hasattr(self._adapter.packundo, 'deleteObject'):
             interface.alsoProvides(self, ZODB.interfaces.IExternalGC)
 
-        self._tpc_phase = _NotInTPCState(self)
+        self._tpc_phase = NotInTransaction(self)
 
     def __repr__(self):
         return "<%s at %x keep_history=%s phase=%r cache=%r>" % (
@@ -708,7 +708,7 @@ class RelStorage(UndoLogCompatible,
             # tid is a committed transaction we will restore.
             # The allowed actions are carefully prescribed.
             # This argument is specified by IStorageRestoreable
-            self._tpc_phase = _RestoreBeginTPCState(self._tpc_phase, tid, status)
+            self._tpc_phase = Restore(self._tpc_phase, tid, status)
 
     def tpc_transaction(self):
         return self._tpc_phase.transaction
@@ -1243,677 +1243,3 @@ def _zlibstorage_new_instance(self):
         if v is not None:
             setattr(new_self, name, v)
     return new_self
-
-
-class _AbstractTPCState(object):
-    """
-    ABC for things a phase of TPC should be able to do.
-    """
-
-    # The transaction
-    transaction = None
-    prepared_txn = None
-
-    aborted = False
-    storage = None
-
-    # - store
-    # - restore/restoreBlob
-    # - deleteObject
-    # - undo
-
-    # should raise ReadOnlyError if the storage is read only.
-
-    # - tpc_vote should raise StorageTransactionError
-
-    # Because entering tpc_begin wasn't allowed if the storage was
-    # read only, this needs to happen in the "not in transaction"
-    # state.
-
-    def tpc_finish(self, transaction, f=None):
-        # For the sake of some ZODB tests, we need to implement this everywhere,
-        # even if it's not actually usable, and the first thing it needs to
-        # do is check the transaction.
-        if transaction is not self.transaction:
-            raise StorageTransactionError('tpc_finish called with wrong transaction')
-        raise NotImplementedError("tpc_finish not allowed in this state.")
-
-    def tpc_abort(self, transaction):
-        if transaction is not self.transaction:
-            return self
-
-        # the lock is held here
-        try:
-            try:
-                self.storage._rollback_load_connection()
-                if self.storage._store_cursor is not None:
-                    self.storage._adapter.txncontrol.abort(
-                        self.storage._store_conn,
-                        self.storage._store_cursor,
-                        self.prepared_txn)
-                    self.storage._adapter.locker.release_commit_lock(self.storage._store_cursor)
-                if self.storage.blobhelper is not None:
-                    self.storage.blobhelper.abort()
-            finally:
-                self._clear_temp()
-        finally:
-            self.aborted = True
-            self.storage._commit_lock.release()
-        return _NotInTPCState(self.storage)
-
-    def _clear_temp(self):
-        # Clear all attributes used for transaction commit.
-        # It is assumed that self._lock.acquire was called before this
-        # method was called.
-        self.storage._cache.clear_temp()
-        blobhelper = self.storage.blobhelper
-        if blobhelper is not None:
-            blobhelper.clear_temp()
-
-class _NotInTPCState(_AbstractTPCState):
-
-    __slots__ = ('is_read_only',)
-
-    def __init__(self, storage):
-        self.is_read_only = storage._is_read_only
-
-    def tpc_abort(self, *args, **kwargs): # pylint:disable=arguments-differ,unused-argument
-        # Nothing to do
-        return self
-
-    def _no_transaction(self, *args, **kwargs):
-        raise StorageTransactionError("No transaction in progress")
-
-    tpc_finish = tpc_vote = _no_transaction
-
-    def store(self, *_args, **_kwargs):
-        if self.is_read_only:
-            raise ReadOnlyError()
-        self._no_transaction()
-
-    restore = deleteObject = undo = restoreBlob = store
-
-    # This object appears to be false.
-    def __bool__(self):
-        return False
-    __nonzero__ = __bool__
-
-class _AbstractBeginTPCState(_AbstractTPCState):
-    """
-    The phase we enter after ``tpc_begin`` has been called.
-    """
-
-    # (user, description, extension) from the transaction.
-    ude = None
-
-    # max_stored_oid is the highest OID stored by the current
-    # transaction
-    max_stored_oid = 0
-
-    # required_tids: {oid_int: tid_int}; confirms that certain objects
-    # have not changed at commit. May be a BTree
-    required_tids = ()
-
-    def __init__(self, storage, transaction):
-        self.storage = storage
-        self.transaction = transaction
-        storage._lock.acquire()
-        try:
-            storage._lock.release()
-            storage._commit_lock.acquire()
-            storage._lock.acquire()
-            storage.transaction = transaction
-
-            user = _to_utf8(transaction.user)
-            desc = _to_utf8(transaction.description)
-            ext = transaction.extension
-
-            if ext:
-                ext = dumps(ext, 1)
-            else:
-                ext = b""
-            self.ude = user, desc, ext
-
-            storage._restart_store()
-            storage._cache.tpc_begin()
-        finally:
-            storage._lock.release()
-
-    def _tpc_vote_factory(self, state):
-        "Return the object with the enter() method."
-        raise NotImplementedError
-
-    def tpc_vote(self, transaction):
-        with self.storage._lock:
-            if transaction is not self.transaction:
-                raise StorageTransactionError(
-                    "tpc_vote called with wrong transaction")
-
-            next_phase = self._tpc_vote_factory(self)
-            next_phase.enter()
-            return next_phase
-
-class _RestoreBeginTPCState(object):
-    # Wraps another begin state and adds the methods needed to
-    # restore or commit to a particular tid.
-
-    # batcher: An object that accumulates store operations
-    # so they can be executed in batch (to minimize latency).
-    batcher = None
-
-    # _batcher_row_limit: The number of rows to queue before
-    # calling the database.
-    batcher_row_limit = 100
-
-    # Methods from the underlying begin implementation we need to
-    # expose.
-    _COPY_ATTRS = (
-        'store',
-        'checkCurrentSerialInTransaction',
-        'deletObject',
-        'undo',
-        'tpc_vote',
-    )
-
-    def __init__(self, begin_state, committing_tid, status):
-        # This is an extension we use for copyTransactionsFrom;
-        # it is not part of the IStorage API.
-        assert committing_tid is not None
-        self.wrapping = begin_state
-
-        # hold the commit lock and add the transaction now
-        storage = begin_state.storage
-        adapter = storage._adapter
-        cursor = storage._store_cursor
-        packed = (status == 'p')
-        adapter.locker.hold_commit_lock(cursor, ensure_current=True)
-        tid_int = bytes8_to_int64(committing_tid)
-        user, desc, ext = begin_state.ude
-        try:
-            adapter.txncontrol.add_transaction(
-                cursor, tid_int, user, desc, ext, packed)
-        except:
-            storage._drop_store_connection()
-            raise
-        committing_tid_lock = _DatabaseLockedForTid(committing_tid, tid_int, adapter)
-        # This is now only used for restore()
-        self.batcher = adapter.mover.make_batcher(
-            storage._store_cursor,
-            self.batcher_row_limit)
-
-        for name in self._COPY_ATTRS:
-            try:
-                meth = getattr(begin_state, name)
-            except AttributeError:
-                continue
-            else:
-                setattr(self, name, meth)
-
-        orig_factory = begin_state._tpc_vote_factory
-        def tpc_vote_factory(state):
-            vote_state = orig_factory(state)
-            vote_state.committing_tid_lock = committing_tid_lock
-
-            orig_flush = vote_state._flush_temps_to_db
-            def flush(cursor):
-                orig_flush(cursor)
-                self.batcher.flush()
-            vote_state._flush_temps_to_db = flush
-            return vote_state
-
-        begin_state._tpc_vote_factory = tpc_vote_factory
-
-    def restore(self, oid, this_tid, data, prev_txn, transaction):
-        # Similar to store() (see comments in FileStorage.restore for
-        # some differences), but used for importing transactions.
-        # Note that *data* can be None.
-        # The *prev_txn* "backpointer" optimization/hint is ignored.
-        #
-        # pylint:disable=unused-argument
-        state = self.wrapping
-        if transaction is not state.transaction:
-            raise StorageTransactionError(self, transaction)
-
-        adapter = state.storage._adapter
-        cursor = state.storage._store_cursor
-        assert cursor is not None
-        oid_int = bytes8_to_int64(oid)
-        tid_int = bytes8_to_int64(this_tid)
-
-        with state.storage._lock:
-            state.max_stored_oid = max(state.max_stored_oid, oid_int)
-            # Save the `data`.  Note that `data` can be None.
-            # Note also that this doesn't go through the cache.
-
-            # TODO: Make it go through the cache, or at least the same
-            # sort of queing thing, so that we can do a bulk COPY?
-            # This complicates restoreBlob().
-            adapter.mover.restore(
-                cursor, self.batcher, oid_int, tid_int, data)
-
-    def restoreBlob(self, oid, serial, data, blobfilename, prev_txn, txn):
-        self.restore(oid, serial, data, prev_txn, txn)
-        state = self.wrapping
-        with state.storage._lock:
-            # Restoring the entry for the blob MAY have used the batcher, and
-            # we're going to want to foreign-key off of that data when
-            # we add blob chunks (since we skip the temp tables).
-            # Ideally, we wouldn't need to flush the batcher here
-            # (we'd prefer having DEFERRABLE INITIALLY DEFERRED FK
-            # constraints, but as-of 8.0 MySQL doesn't support that.)
-            self.batcher.flush()
-            cursor = state.storage._store_cursor
-            state.storage.blobhelper.restoreBlob(cursor, oid, serial, blobfilename)
-
-
-class _AbstractStoreBeginTPCState(_AbstractBeginTPCState):
-    # pylint:disable=abstract-method
-
-    def store(self, oid, previous_tid, data, transaction):
-        # Called by Connection.commit(), after tpc_begin has been called.
-        if transaction is not self.transaction:
-            raise StorageTransactionError(self, transaction)
-
-        #adapter = self._adapter
-        cache = self.storage._cache
-        cursor = self.storage._store_cursor
-        assert cursor is not None
-        oid_int = bytes8_to_int64(oid)
-        if previous_tid:
-            # previous_tid is the tid of the state that the
-            # object was loaded from.
-
-            # XXX PY3: ZODB.tests.IteratorStorage passes a str (non-bytes) value for oid
-            prev_tid_int = bytes8_to_int64(
-                previous_tid
-                if isinstance(previous_tid, bytes)
-                else previous_tid.encode('ascii')
-            )
-        else:
-            prev_tid_int = 0
-
-        with self.storage._lock:
-            self.max_stored_oid = max(self.max_stored_oid, oid_int)
-            # Save the data locally in a temporary place. Later, closer to commit time,
-            # we'll send it all over at once. This lets us do things like use
-            # COPY in postgres.
-            cache.store_temp(oid_int, data, prev_tid_int)
-            return None
-
-    def checkCurrentSerialInTransaction(self, oid, required_tid, transaction):
-        if transaction is not self.transaction:
-            raise StorageTransactionError(self, transaction)
-
-        _, committed_tid = self.storage.load(oid)
-        if committed_tid != required_tid:
-            raise ReadConflictError(
-                oid=oid, serials=(committed_tid, required_tid))
-
-        required_tid_int = bytes8_to_int64(required_tid)
-        oid_int = bytes8_to_int64(oid)
-
-        # If this transaction already specified a different serial for
-        # this oid, the transaction conflicts with itself.
-        required_tids = self.required_tids
-        if not required_tids:
-            required_tids = self.required_tids = OID_TID_MAP_TYPE()
-
-        previous_serial_int = required_tids.get(oid_int, required_tid_int)
-        if previous_serial_int != required_tid_int:
-            raise ReadConflictError(
-                oid=oid,
-                serials=(int64_to_8bytes(previous_serial_int),
-                         required_tid))
-        required_tids[oid_int] = required_tid_int
-
-class _HistoryFreeBeginTPCState(_AbstractStoreBeginTPCState):
-
-    def deleteObject(self, oid, oldserial, transaction):
-        # This method is only expected to be called from zc.zodbdgc
-        # currently, or from ZODB/tests/IExternalGC.test
-
-        # This is called in a phase of two-phase-commit (tpc).
-        # This means we have a transaction, and that we are holding
-        # the commit lock as well as the regular lock.
-        # RelStorage native pack uses a separate pack lock, but
-        # unfortunately there's no way to not hold the commit lock;
-        # however, the transactions are very short.
-        if transaction is not self.transaction: # pragma: no cover
-            raise StorageTransactionError(self, transaction)
-
-        # We don't worry about anything in self._cache because
-        # by definition we are deleting objects that were
-        # not reachable and so shouldn't be in the cache (or if they
-        # were, we'll never ask for them anyway)
-
-        # We delegate the actual operation to the adapter's packundo,
-        # just like native pack
-        cursor = self.storage._store_cursor
-        assert cursor is not None
-        # When this is done, we get a tpc_vote,
-        # and a tpc_finish.
-        # The interface doesn't specify a return value, so for testing
-        # we return the count of rows deleted (should be 1 if successful)
-        return self.storage._adapter.packundo.deleteObject(cursor, oid, oldserial)
-
-class _DatabaseLockedForTid(object):
-
-    def __init__(self, tid, tid_int, adapter):
-        self.tid = tid
-        self.tid_int = tid_int
-        self._adapter = adapter
-
-    def release_commit_lock(self, cursor):
-        self._adapter.locker.release_commit_lock(cursor)
-
-def _prepare_tid_and_lock_database(adapter, cursor, ude):
-    """
-    Returns an _DatabaseLockedForTid.
-    """
-    # TODO: Stop doing this here; go to row-level locking.
-    adapter.locker.hold_commit_lock(cursor, ensure_current=True)
-    user, desc, ext = ude
-
-    # Choose a transaction ID.
-    # Base the transaction ID on the current time,
-    # but ensure that the tid of this transaction
-    # is greater than any existing tid.
-    # TODO: Stop allocating this here. Defer until tpc_finish.
-    last_tid = adapter.txncontrol.get_tid(cursor)
-    now = time.time()
-    stamp = TimeStamp(*(time.gmtime(now)[:5] + (now % 60,)))
-    stamp = stamp.laterThan(TimeStamp(int64_to_8bytes(last_tid)))
-    tid = stamp.raw()
-
-    tid_int = bytes8_to_int64(tid)
-    adapter.txncontrol.add_transaction(cursor, tid_int, user, desc, ext)
-    return _DatabaseLockedForTid(tid, tid_int, adapter)
-
-class _HistoryPreservingBeginTPCState(_AbstractStoreBeginTPCState):
-
-    # Stored in their 8 byte form
-    undone_oids = ()
-
-    # If we use undo(), we have to allocate a TID, which means
-    # we have to lock the database. Not cool.
-    committing_tid_lock = None
-
-    def undo(self, transaction_id, transaction):
-        # Typically if this is called, the store/restore methods will *not* be
-        # called, but there's not a strict guarantee about that.
-        if transaction is not self.transaction:
-            raise StorageTransactionError(self, transaction)
-
-        # Unlike most places, transaction_id is the base 64 encoding
-        # of an 8 byte tid
-
-        undo_tid = base64_decodebytes(transaction_id + b'\n') # pylint:disable=deprecated-method
-        assert len(undo_tid) == 8
-        undo_tid_int = bytes8_to_int64(undo_tid)
-
-        with self.storage._lock:
-            adapter = self.storage._adapter
-            cursor = self.storage._store_cursor
-            assert cursor is not None
-
-            adapter.locker.hold_pack_lock(cursor)
-            try:
-                adapter.packundo.verify_undoable(cursor, undo_tid_int)
-                if self.committing_tid_lock is None:
-                    # Note that _prepare_tid acquires the commit lock.
-                    # The commit lock must be acquired after the pack lock
-                    # because the database adapters also acquire in that
-                    # order during packing.
-                    adapter = self.storage._adapter
-                    cursor = self.storage._store_cursor
-                    tid_lock = _prepare_tid_and_lock_database(adapter, cursor, self.ude)
-                    self.committing_tid_lock = tid_lock
-
-                self_tid_int = self.committing_tid_lock.tid_int
-                copied = adapter.packundo.undo(
-                    cursor, undo_tid_int, self_tid_int)
-                oids = [int64_to_8bytes(oid_int) for oid_int, _ in copied]
-
-                # Update the current object pointers immediately, so that
-                # subsequent undo operations within this transaction will see
-                # the new current objects.
-                adapter.mover.update_current(cursor, self_tid_int)
-
-                if self.storage.blobhelper is not None:
-                    self.storage.blobhelper.copy_undone(copied,
-                                                        self.committing_tid_lock.tid)
-
-                if not self.undone_oids:
-                    self.undone_oids = set()
-                self.undone_oids.update(oids)
-            finally:
-                adapter.locker.release_pack_lock(cursor)
-
-class _AbstractVoteTPCState(_AbstractTPCState):
-    """
-    The state we're in following ``tpc_vote``.
-
-    Unlike the begin states, you *must* explicitly call :meth:`enter`
-    on this object after it is constructed.
-
-    """
-
-    # The byte representation of the TID we've picked to commit.
-    committing_tid_lock = None
-
-    prepared_txn = None
-
-    def __init__(self, begin_state, committing_tid_lock=None):
-        # If committing_tid is passed to this method, it means
-        # the database has already been locked and the TID is
-        # locked in.
-        self.storage = begin_state.storage
-        self.transaction = begin_state.transaction
-        self.required_tids = begin_state.required_tids
-        self.max_stored_oid = begin_state.max_stored_oid
-        self.ude = begin_state.ude
-        self.committing_tid_lock = committing_tid_lock
-        self.resolved_oids = set()
-
-    def enter(self):
-        resolved_in_vote = self.__vote()
-        self.resolved_oids.update(resolved_in_vote)
-
-    def _flush_temps_to_db(self, cursor):
-        mover = self.storage._adapter.mover
-        mover.store_temps(cursor, self.storage._cache.temp_objects)
-
-    def __vote(self):
-        """
-        Prepare the transaction for final commit.
-
-        Takes the exclusive database commit lock.
-        """
-        # This method initiates a two-phase commit process,
-        # saving the name of the prepared transaction in self._prepared_txn.
-
-        # It is assumed that self._lock.acquire was called before this
-        # method was called.
-        cursor = self.storage._store_cursor
-        assert cursor is not None
-        conn = self.storage._store_conn
-        adapter = self.storage._adapter
-        mover = adapter.mover
-
-        # execute all remaining batch store operations
-        self._flush_temps_to_db(cursor)
-
-        # Reserve all OIDs used by this transaction
-        if self.max_stored_oid > self.storage._max_new_oid:
-            adapter.oidallocator.set_min_oid(
-                cursor, self.max_stored_oid + 1)
-
-        # Here's where we take the global commit lock, and
-        # allocate the next available transaction id, storing it
-        # into history-preserving DBs. But if someone passed us
-        # one, then it must already be in the DB, and the lock must
-        # already be held.
-        if not self.committing_tid_lock:
-            self.__prepare_tid()
-
-        committing_tid_int = self.committing_tid_lock.tid_int
-
-        # XXX: When we stop allocating the TID early, we need to move
-        # 'move_from_temp' (hf and hp) and 'update_current' (hp only)
-        # later, down to tpc_finish. With everything safely locked,
-        # that shouldn't fail. But we still need to check serials and
-        # resolve conflicts here. When we do that, we need to lock all
-        # the rows involved for update, in order (to avoid deadlocks),
-        # including the prev_tid rows (hp only) and current committed
-        # rows (both) for everything that conflicts.
-
-        # Check the things registered by Connection.readCurrent()
-        if self.required_tids:
-            oid_ints = self.required_tids.keys()
-            current = mover.current_object_tids(cursor, oid_ints)
-            for oid_int, expect_tid_int in self.required_tids.items():
-                actual_tid_int = current.get(oid_int, 0)
-                if actual_tid_int != expect_tid_int:
-                    raise ReadConflictError(
-                        oid=int64_to_8bytes(oid_int),
-                        serials=(int64_to_8bytes(actual_tid_int),
-                                 int64_to_8bytes(expect_tid_int)))
-
-        resolved_serials = self.__finish_store(committing_tid_int)
-        self.storage._adapter.mover.update_current(cursor, committing_tid_int)
-        self.prepared_txn = self.storage._adapter.txncontrol.commit_phase1(
-            conn, cursor, committing_tid_int)
-
-        if self.storage.blobhelper is not None:
-            self.storage.blobhelper.vote(self.committing_tid_lock.tid)
-
-        # New storage protocol
-        return resolved_serials
-
-    def __prepare_tid(self):
-        """
-        Choose a tid for the current transaction, and exclusively lock
-        the database commit lock.
-
-        This should be done as late in the commit as possible, since
-        it must hold an exclusive commit lock.
-        """
-        adapter = self.storage._adapter
-        cursor = self.storage._store_cursor
-        # TODO: Stop doing this here; go to row-level locking.
-        lock = _prepare_tid_and_lock_database(adapter, cursor, self.ude)
-        self.committing_tid_lock = lock
-        return lock
-
-    def __finish_store(self, committing_tid_int):
-        """
-        Move stored objects from the temporary table to final storage.
-
-        Returns a sequence of OIDs that were resolved to be received
-        by Connection._handle_serial().
-        """
-        # pylint:disable=too-many-locals
-        cursor = self.storage._store_cursor
-        adapter = self.storage._adapter
-        cache = self.storage._cache
-        tryToResolveConflict = self.storage.tryToResolveConflict
-
-        # Detect conflicting changes.
-        # Try to resolve the conflicts.
-        resolved = set()  # a set of OIDs
-        # In the past, we didn't load all conflicts from the DB at once,
-        # just one at a time. This was because we also fetched the state data
-        # from the DB, and it could be large. But now we use the state we have in
-        # our local temp cache, so memory concerns are gone.
-        conflicts = adapter.mover.detect_conflict(cursor)
-        if conflicts:
-            log.debug("Attempting to resolve %d conflicts", len(conflicts))
-
-        for conflict in conflicts:
-            oid_int, committed_tid_int, tid_this_txn_saw_int = conflict
-            state_from_this_txn = cache.read_temp(oid_int)
-            oid = int64_to_8bytes(oid_int)
-            prev_tid = int64_to_8bytes(committed_tid_int)
-            serial = int64_to_8bytes(tid_this_txn_saw_int)
-
-            resolved_state = tryToResolveConflict(oid, prev_tid, serial, state_from_this_txn)
-            if resolved_state is None:
-                # unresolvable; kill the whole transaction
-                raise ConflictError(
-                    oid=oid, serials=(prev_tid, serial), data=state_from_this_txn)
-
-            # resolved
-            state_from_this_txn = resolved_state
-            # TODO: Make this use the bulk methods so we can use COPY.
-            adapter.mover.replace_temp(
-                cursor, oid_int, committed_tid_int, state_from_this_txn)
-            resolved.add(oid)
-            cache.store_temp(oid_int, state_from_this_txn)
-
-        # Move the new states into the permanent table
-        if self.storage.blobhelper is not None:
-            txn_has_blobs = self.storage.blobhelper.txn_has_blobs
-        else:
-            txn_has_blobs = False
-
-        # This returns the OID ints stored, but we don't use them here
-        adapter.mover.move_from_temp(cursor, committing_tid_int, txn_has_blobs)
-
-        return resolved
-
-    def tpc_finish(self, transaction, f=None):
-        with self.storage._lock:
-            if transaction is not self.transaction:
-                raise StorageTransactionError(
-                    "tpc_finish called with wrong transaction")
-            assert self.committing_tid_lock is not None, self
-            try:
-                try:
-                    if f is not None:
-                        f(self.committing_tid_lock.tid)
-                    next_phase = _tpc_finish_factory(self)
-                    return next_phase, self.committing_tid_lock.tid
-                finally:
-                    self._clear_temp()
-            finally:
-                self.storage._commit_lock.release()
-
-
-class _HistoryFreeVoteTPCState(_AbstractVoteTPCState):
-    pass
-
-_HistoryFreeBeginTPCState._tpc_vote_factory = _HistoryFreeVoteTPCState
-
-class _HistoryPreservingVoteTPCState(_AbstractVoteTPCState):
-
-    def __init__(self, begin_state):
-        assert isinstance(begin_state, _HistoryPreservingBeginTPCState)
-         # Using undo() requires a new TID, so if we had already begun
-        # a transaction by locking the database and allocating a TID,
-        # we must preserve that.
-        super(_HistoryPreservingVoteTPCState, self).__init__(begin_state,
-                                                             begin_state.committing_tid_lock)
-        self.resolved_oids.update(begin_state.undone_oids)
-
-_HistoryPreservingBeginTPCState._tpc_vote_factory = _HistoryPreservingVoteTPCState
-
-def _tpc_finish_factory(vote_state):
-    """
-    The state we enter with tpc_finish.
-
-    This is transient; once we successfully enter this state, we immediately return
-    to the not-in-transaction state.
-    """
-    # It is assumed that self._lock.acquire was called before this
-    # method was called.
-    vote_state.storage._rollback_load_connection()
-    txn = vote_state.prepared_txn
-    assert txn is not None
-    vote_state.storage._adapter.txncontrol.commit_phase2(
-        vote_state.storage._store_conn,
-        vote_state.storage._store_cursor,
-        txn)
-    vote_state.committing_tid_lock.release_commit_lock(vote_state.storage._store_cursor)
-    vote_state.storage._cache.after_tpc_finish(vote_state.committing_tid_lock.tid)
-
-    return _NotInTPCState(vote_state.storage)
