+++ src/relstorage/storage/tpc/vote.py
@@ -1,0 +1,280 @@
+##############################################################################
+#
+# Copyright (c) 2019 Zope Foundation and Contributors.
+# All Rights Reserved.
+#
+# This software is subject to the provisions of the Zope Public License,
+# Version 2.1 (ZPL).  A copy of the ZPL should accompany this distribution.
+# THIS SOFTWARE IS PROVIDED "AS IS" AND ANY AND ALL EXPRESS OR IMPLIED
+# WARRANTIES ARE DISCLAIMED, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+# WARRANTIES OF TITLE, MERCHANTABILITY, AGAINST INFRINGEMENT, AND FITNESS
+# FOR A PARTICULAR PURPOSE.
+#
+##############################################################################
+"""
+States related to TPC voting.
+
+The implementation of locking the database for commit, flushing
+temporary objects to the database, and moving them to their final locations,
+live here.
+"""
+from __future__ import absolute_import
+from __future__ import print_function
+
+import time
+
+from persistent.timestamp import TimeStamp
+from ZODB.POSException import ConflictError
+from ZODB.POSException import ReadConflictError
+from ZODB.POSException import StorageTransactionError
+from ZODB.utils import p64 as int64_to_8bytes
+from ZODB.utils import u64 as bytes8_to_int64
+
+
+from . import AbstractTPCState
+from .finish import Finish
+
+logger = __import__('logging').getLogger(__name__)
+
+class DatabaseLockedForTid(object):
+
+    @classmethod
+    def lock_database_for_next_tid(cls, cursor, adapter, ude):
+        # TODO: Stop doing this here; go to row-level locking.
+        adapter.locker.hold_commit_lock(cursor, ensure_current=True)
+        user, desc, ext = ude
+
+        # Choose a transaction ID.
+        # Base the transaction ID on the current time,
+        # but ensure that the tid of this transaction
+        # is greater than any existing tid.
+        # TODO: Stop allocating this here. Defer until tpc_finish.
+        last_tid = adapter.txncontrol.get_tid(cursor)
+        now = time.time()
+        stamp = TimeStamp(*(time.gmtime(now)[:5] + (now % 60,)))
+        stamp = stamp.laterThan(TimeStamp(int64_to_8bytes(last_tid)))
+        tid = stamp.raw()
+
+        tid_int = bytes8_to_int64(tid)
+        adapter.txncontrol.add_transaction(cursor, tid_int, user, desc, ext)
+        return cls(tid, tid_int, adapter)
+
+    @classmethod
+    def lock_database_for_given_tid(cls, tid, tid_is_packed,
+                                    cursor, adapter, ude):
+        adapter.locker.hold_commit_lock(cursor, ensure_current=True)
+        tid_int = bytes8_to_int64(tid)
+        user, desc, ext = ude
+        adapter.txncontrol.add_transaction(
+            cursor, tid_int, user, desc, ext, tid_is_packed)
+        return cls(tid, tid_int, adapter)
+
+    def __init__(self, tid, tid_int, adapter):
+        self.tid = tid
+        self.tid_int = tid_int
+        self.release_commit_lock = adapter.locker.release_commit_lock
+
+
+class AbstractVote(AbstractTPCState):
+    """
+    The state we're in following ``tpc_vote``.
+
+    Unlike the begin states, you *must* explicitly call :meth:`enter`
+    on this object after it is constructed.
+
+    """
+
+    # The byte representation of the TID we've picked to commit.
+    committing_tid_lock = None
+
+    prepared_txn = None
+
+    def __init__(self, begin_state, committing_tid_lock=None):
+        # If committing_tid is passed to this method, it means
+        # the database has already been locked and the TID is
+        # locked in.
+        self.storage = begin_state.storage
+        self.transaction = begin_state.transaction
+        self.required_tids = begin_state.required_tids
+        self.max_stored_oid = begin_state.max_stored_oid
+        self.ude = begin_state.ude
+        self.committing_tid_lock = committing_tid_lock
+        self.resolved_oids = set()
+
+    def enter(self):
+        resolved_in_vote = self.__vote()
+        self.resolved_oids.update(resolved_in_vote)
+
+    def _flush_temps_to_db(self, cursor):
+        mover = self.storage._adapter.mover
+        mover.store_temps(cursor, self.storage._cache.temp_objects)
+
+    def __vote(self):
+        """
+        Prepare the transaction for final commit.
+
+        Takes the exclusive database commit lock.
+        """
+        # This method initiates a two-phase commit process,
+        # saving the name of the prepared transaction in self._prepared_txn.
+
+        # It is assumed that self._lock.acquire was called before this
+        # method was called.
+        cursor = self.storage._store_cursor
+        assert cursor is not None
+        conn = self.storage._store_conn
+        adapter = self.storage._adapter
+        mover = adapter.mover
+
+        # execute all remaining batch store operations
+        self._flush_temps_to_db(cursor)
+
+        # Reserve all OIDs used by this transaction
+        if self.max_stored_oid > self.storage._max_new_oid:
+            adapter.oidallocator.set_min_oid(
+                cursor, self.max_stored_oid + 1)
+
+        # Here's where we take the global commit lock, and
+        # allocate the next available transaction id, storing it
+        # into history-preserving DBs. But if someone passed us
+        # one, then it must already be in the DB, and the lock must
+        # already be held.
+        if not self.committing_tid_lock:
+            self.__prepare_tid()
+
+        committing_tid_int = self.committing_tid_lock.tid_int
+
+        # XXX: When we stop allocating the TID early, we need to move
+        # 'move_from_temp' (hf and hp) and 'update_current' (hp only)
+        # later, down to tpc_finish. With everything safely locked,
+        # that shouldn't fail. But we still need to check serials and
+        # resolve conflicts here. When we do that, we need to lock all
+        # the rows involved for update, in order (to avoid deadlocks),
+        # including the prev_tid rows (hp only) and current committed
+        # rows (both) for everything that conflicts.
+
+        # Check the things registered by Connection.readCurrent()
+        if self.required_tids:
+            oid_ints = self.required_tids.keys()
+            current = mover.current_object_tids(cursor, oid_ints)
+            for oid_int, expect_tid_int in self.required_tids.items():
+                actual_tid_int = current.get(oid_int, 0)
+                if actual_tid_int != expect_tid_int:
+                    raise ReadConflictError(
+                        oid=int64_to_8bytes(oid_int),
+                        serials=(int64_to_8bytes(actual_tid_int),
+                                 int64_to_8bytes(expect_tid_int)))
+
+        resolved_serials = self.__finish_store(committing_tid_int)
+        self.storage._adapter.mover.update_current(cursor, committing_tid_int)
+        self.prepared_txn = self.storage._adapter.txncontrol.commit_phase1(
+            conn, cursor, committing_tid_int)
+
+        if self.storage.blobhelper is not None:
+            self.storage.blobhelper.vote(self.committing_tid_lock.tid)
+
+        # New storage protocol
+        return resolved_serials
+
+    def __prepare_tid(self):
+        """
+        Choose a tid for the current transaction, and exclusively lock
+        the database commit lock.
+
+        This should be done as late in the commit as possible, since
+        it must hold an exclusive commit lock.
+        """
+        adapter = self.storage._adapter
+        cursor = self.storage._store_cursor
+        # TODO: Stop doing this here; go to row-level locking.
+        lock = DatabaseLockedForTid.lock_database_for_next_tid(cursor, adapter, self.ude)
+        self.committing_tid_lock = lock
+        return lock
+
+    def __finish_store(self, committing_tid_int):
+        """
+        Move stored objects from the temporary table to final storage.
+
+        Returns a sequence of OIDs that were resolved to be received
+        by Connection._handle_serial().
+        """
+        # pylint:disable=too-many-locals
+        cursor = self.storage._store_cursor
+        adapter = self.storage._adapter
+        cache = self.storage._cache
+        tryToResolveConflict = self.storage.tryToResolveConflict
+
+        # Detect conflicting changes.
+        # Try to resolve the conflicts.
+        resolved = set()  # a set of OIDs
+        # In the past, we didn't load all conflicts from the DB at once,
+        # just one at a time. This was because we also fetched the state data
+        # from the DB, and it could be large. But now we use the state we have in
+        # our local temp cache, so memory concerns are gone.
+        conflicts = adapter.mover.detect_conflict(cursor)
+        if conflicts:
+            logger.debug("Attempting to resolve %d conflicts", len(conflicts))
+
+        for conflict in conflicts:
+            oid_int, committed_tid_int, tid_this_txn_saw_int = conflict
+            state_from_this_txn = cache.read_temp(oid_int)
+            oid = int64_to_8bytes(oid_int)
+            prev_tid = int64_to_8bytes(committed_tid_int)
+            serial = int64_to_8bytes(tid_this_txn_saw_int)
+
+            resolved_state = tryToResolveConflict(oid, prev_tid, serial, state_from_this_txn)
+            if resolved_state is None:
+                # unresolvable; kill the whole transaction
+                raise ConflictError(
+                    oid=oid, serials=(prev_tid, serial), data=state_from_this_txn)
+
+            # resolved
+            state_from_this_txn = resolved_state
+            # TODO: Make this use the bulk methods so we can use COPY.
+            adapter.mover.replace_temp(
+                cursor, oid_int, committed_tid_int, state_from_this_txn)
+            resolved.add(oid)
+            cache.store_temp(oid_int, state_from_this_txn)
+
+        # Move the new states into the permanent table
+        if self.storage.blobhelper is not None:
+            txn_has_blobs = self.storage.blobhelper.txn_has_blobs
+        else:
+            txn_has_blobs = False
+
+        # This returns the OID ints stored, but we don't use them here
+        adapter.mover.move_from_temp(cursor, committing_tid_int, txn_has_blobs)
+
+        return resolved
+
+    def tpc_finish(self, transaction, f=None):
+        with self.storage._lock:
+            if transaction is not self.transaction:
+                raise StorageTransactionError(
+                    "tpc_finish called with wrong transaction")
+            assert self.committing_tid_lock is not None, self
+            try:
+                try:
+                    if f is not None:
+                        f(self.committing_tid_lock.tid)
+                    next_phase = Finish(self)
+                    return next_phase, self.committing_tid_lock.tid
+                finally:
+                    self._clear_temp()
+            finally:
+                self.storage._commit_lock.release()
+
+
+class HistoryFree(AbstractVote):
+    pass
+
+
+class HistoryPreserving(AbstractVote):
+
+    def __init__(self, begin_state):
+         # Using undo() requires a new TID, so if we had already begun
+        # a transaction by locking the database and allocating a TID,
+        # we must preserve that.
+        super(HistoryPreserving, self).__init__(begin_state,
+                                                begin_state.committing_tid_lock)
+        self.resolved_oids.update(begin_state.undone_oids)
