+++ src/relstorage/storage/transaction_iterator.py
